refactor(test): break up test, pass capabilities on per-test basis
