+++ test-suite/src/hover.rs
@@ -1,0 +1,149 @@
+#[cfg(test)]
+mod test {
+    use std::{num::NonZeroU32, time::Duration};
+
+    use crate::test_helpers::cargo_dot_toml;
+    use lspresso_shot::{
+        lspresso_shot, test_hover,
+        types::{ServerStartType, TestCase, TestFile},
+    };
+    use test_server::{send_capabiltiies, send_response_num};
+
+    use lsp_types::{
+        Hover, HoverContents, HoverOptions, HoverProviderCapability, MarkupContent, MarkupKind,
+        Position, Range, ServerCapabilities, WorkDoneProgressOptions,
+    };
+    use rstest::rstest;
+
+    fn hover_capabilities_simple() -> ServerCapabilities {
+        ServerCapabilities {
+            hover_provider: Some(HoverProviderCapability::Simple(true)),
+            ..ServerCapabilities::default()
+        }
+    }
+
+    // TODO: Implement a mock progress-style hover response, and switch the dispatch
+    // based on that
+    #[allow(dead_code)]
+    fn hover_capabilities_progress() -> ServerCapabilities {
+        ServerCapabilities {
+            hover_provider: Some(HoverProviderCapability::Options(HoverOptions {
+                work_done_progress_options: WorkDoneProgressOptions {
+                    work_done_progress: Some(true),
+                },
+            })),
+            ..ServerCapabilities::default()
+        }
+    }
+
+    #[rstest]
+    fn test_server_hover_simple(#[values(0, 1, 2, 3, 4, 5)] response_num: u32) {
+        use test_server::get_dummy_server_path;
+
+        let resp = test_server::responses::get_hover_response(response_num).unwrap();
+        let source_file = TestFile::new(test_server::get_source_path(), "");
+        let test_case = TestCase::new(get_dummy_server_path(), source_file)
+            .cursor_pos(Some(Position::default()))
+            .cleanup(false);
+
+        let test_case_root = test_case
+            .get_lspresso_dir()
+            .expect("Failed to get test case's root directory");
+        send_response_num(response_num, &test_case_root).expect("Failed to send response num");
+        send_capabiltiies(&hover_capabilities_simple(), &test_case_root)
+            .expect("Failed to send capabilities");
+
+        lspresso_shot!(test_hover(test_case, resp));
+    }
+
+    #[test]
+    fn rust_analyzer_hover() {
+        let source_file = TestFile::new(
+            "src/main.rs",
+            r#"pub fn main() {
+    println!("Hello, world!");
+}"#,
+        );
+        let test_case = TestCase::new("rust-analyzer", source_file)
+            .start_type(ServerStartType::Progress(
+                NonZeroU32::new(1).unwrap(),
+                "rustAnalyzer/Indexing".to_string(),
+            ))
+            .timeout(Duration::from_secs(20))
+            .cleanup(false)
+            .cursor_pos(Some(Position::new(1, 5)))
+            .other_file(cargo_dot_toml());
+
+        lspresso_shot!(test_hover(
+        test_case,
+        Hover {
+            range: Some(Range {
+                start: Position {
+                    line: 1,
+                    character: 4,
+                },
+                end: Position {
+                    line: 1,
+                    character: 11,
+                },
+            }),
+            contents: HoverContents::Markup(MarkupContent {
+                kind: MarkupKind::Markdown,
+                value:
+                "
+```rust
+std::macros
+```
+
+```rust
+macro_rules! println // matched arm #1
+```
+
+---
+
+Prints to the standard output, with a newline.
+
+On all platforms, the newline is the LINE FEED character (`\\n`/`U+000A`) alone
+(no additional CARRIAGE RETURN (`\\r`/`U+000D`)).
+
+This macro uses the same syntax as [`format`](https://doc.rust-lang.org/stable/alloc/macros/macro.format.html), but writes to the standard output instead.
+See [`std::fmt`] for more information.
+
+The `println!` macro will lock the standard output on each call. If you call
+`println!` within a hot loop, this behavior may be the bottleneck of the loop.
+To avoid this, lock stdout with [`io::stdout().lock`](https://doc.rust-lang.org/stable/std/io/stdio/struct.Stdout.html):
+
+```rust
+use std::io::{stdout, Write};
+
+let mut lock = stdout().lock();
+writeln!(lock, \"hello world\").unwrap();
+```
+
+Use `println!` only for the primary output of your program. Use
+[`eprintln`] instead to print error and progress messages.
+
+See [the formatting documentation in `std::fmt`](https://doc.rust-lang.org/stable/std/std/fmt/index.html)
+for details of the macro argument syntax.
+
+# Panics
+
+Panics if writing to [`io::stdout`] fails.
+
+Writing to non-blocking stdout can cause an error, which will lead
+this macro to panic.
+
+# Examples
+
+```rust
+println!(); // prints just a newline
+println!(\"hello there!\");
+println!(\"format {} arguments\", \"some\");
+let local_variable = \"some\";
+println!(\"format {local_variable} arguments\");
+```".to_string()
+            })
+        }
+    ));
+    }
+}
