+++ test-suite/src/lib.rs
@@ -1,0 +1,10 @@
+pub mod test_helpers;
+
+mod completion;
+mod definition;
+mod diagnostics;
+mod document_symbol;
+mod formatting;
+mod hover;
+mod references;
+mod rename;
